third_party/blink/web_tests/crypto/subtle/aes-cbc/wrap-rsa-expected.txt

Test wrapping an RSA key with AES-CBC.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Importing a key to wrap...
PASS key.algorithm.name is 'RSASSA-PKCS1-v1_5'
Importing a key encryption key...
PASS wrappingKey.algorithm.name is 'AES-CBC'
PASS wrappedKey.toString() is '[object ArrayBuffer]'
PASS successfullyParsed is true

TEST COMPLETE
